refactor: extracts prev and next chapter URL state from BottomToolbar to derived Atom
