+++ src/hooks/useRangeInput.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect, useRef, useState } from 'react'
 
 export const useRangeInput = <T>(
