+++ src/organisms/BottomToolbar/BottomToolbar.tsx
@@ -1,18 +1,16 @@
 'use client'
 
 import { useAtom, useAtomValue } from 'jotai'
-import { useState } from 'react'
 import { hstack, subgrid } from 'styled-system/patterns'
-import { useIsFirstRender } from 'usehooks-ts'
 
 import type { TBook } from '~/db'
-import { buildReaderUrl } from '~/hooks'
 import {
 	currBookAtom,
-	currBookCodeAtom,
 	currChapterAtom,
 	isFirstChapterAtom,
 	isLastChapterAtom,
+	nextChapterUrlAtom,
+	prevChapterUrlAtom,
 } from '~/state'
 
 import { BottomToolbarContainer } from './BottomToolbarContainer'
@@ -21,8 +19,6 @@ import { ReaderNavButton } from './ReaderNavButton'
 import { ReturnFromReferenceFab } from './ReturnFromReferenceFab'
 
 export const BottomToolbar = ({ bookList }: { bookList: TBook[] }) => {
-	const bookCode = useAtomValue(currBookCodeAtom)
-
 	const chapter = useAtomValue(currChapterAtom)
 
 	const currBook = useAtomValue(currBookAtom)
@@ -33,38 +29,9 @@ export const BottomToolbar = ({ bookList }: { bookList: TBook[] }) => {
 
 	const currBookChapterCount = bookList[currBookIndex]?.chapter_count
 
-	const prevBook = bookList[currBookIndex - 1]
+	const prevChapterUrl = useAtomValue(prevChapterUrlAtom)
 
-	const prevBookCode = prevBook?.code
-
-	const prevBookChapterCount = prevBook?.chapter_count
-
-	const nextBookCode = bookList[currBookIndex + 1]?.code
-
-	const [prevChapterUrl, setPrevChapterUrl] = useState('')
-
-	const isFirstRender = useIsFirstRender()
-
-	if (isFirstRender) {
-		setPrevChapterUrl(
-			chapter === 1
-				? buildReaderUrl({
-						bookCode: prevBookCode,
-						chapter: prevBookChapterCount,
-				  })
-				: buildReaderUrl({ bookCode, chapter: chapter - 1 }),
-		)
-	}
-
-	const [nextChapterUrl, setNextChapterUrl] = useState('')
-
-	if (isFirstRender) {
-		setNextChapterUrl(
-			chapter === currBookChapterCount
-				? buildReaderUrl({ bookCode: nextBookCode, chapter: 1 })
-				: buildReaderUrl({ bookCode, chapter: chapter + 1 }),
-		)
-	}
+	const nextChapterUrl = useAtomValue(nextChapterUrlAtom)
 
 	const [isFirstChapterInBible, setIsFirstChapterInBible] =
 		useAtom(isFirstChapterAtom)
@@ -87,7 +54,6 @@ export const BottomToolbar = ({ bookList }: { bookList: TBook[] }) => {
 					pb: 'safe_area_bottom',
 				})}
 			>
-				{prevChapterUrl} | {nextChapterUrl}
 				<ReturnFromReferenceFab />
 				<div
 					className={hstack({
